`setup.py` prepares and setups the yolov3, yolov3tiny architecture and utilies for easy to use more practical code.
`train.py` is able to train a model based on yolo algorithm (tiny or not), picking up through
a variety of options like training from scratch, or transfer learning for specific components
of the model.

`data` directory contains anchors boxes size based on voc and coco dataset, as their objects
are similar to each other and consequently their aspect ratios.

On the other hand, `my_data` is the directory which contains the corresponding information for
fire-smoke dataset.